+++ src/test/kotlin/org/axonframework/intellij/ide/plugin/specifics/ExceptionCasesTests.kt
@@ -36,6 +36,6 @@ class ExceptionCasesTests : AbstractAxonFixtureTestCase() {
         """.trimIndent()
         )
 
-        project.aggregateResolver().getModels()
+        project.aggregateResolver().getMemberForName("text.MyAggregate")
     }
 }