+++ CHANGELOG.md
@@ -4,11 +4,19 @@
 
 ## [0.7.0]
 
+### Fixed
+- [#66] Fix possible race condition between primary and secondary cache computations in annotation resolver
+- [#68] Remove unnecessary catching of errors during resolving a qualifiedName in PsiProcessingUtil
+- [#69] Disable plugin functionality when old versions are used
+- Fix OutOfBoundException in AxonImplicitUsageProvider when using Groovy
+
 ### Added
 
 - [#38] Aggregate structure hierarchy is now shown in model popup
 - [#41] Mark methods annotated with @ResetHandler as used
-- Fix OutOfBoundException in AxonImplicitUsageProvider when using Groovy
+- [#31] Aggregate structure hierarchy is now shown in model popup
+- [#27] Inspection when associationProperty on @SagaEventHandler is missing in the message
+- [#23] Inspection when routingKey is missing in messages
 
 ## [0.6.2]